18. Optimizing the messages

+++ API/SignalR/MessageHub.cs
@@ -30,18 +30,22 @@ namespace API.SignalR
             var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
             await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
 
-            await AddToGroup(groupName);
+            // await AddToGroup(groupName);
+            var group = await AddToGroup(groupName);
+            await Clients.Group(groupName).SendAsync("UpdatedGroup",group);
 
-            //every initial request me hm saraa message send krain gaa 
-            // jo kaa galt ha  end me hm set krain ga isaaa
             var messages = await messageRepository.GetMessageThread(Context.User.GetUsername(), otherUser!);
-
-            await Clients.Group(groupName).SendAsync("ReceiveMessageThread",messages); 
+            
+            await Clients.Caller.SendAsync("ReceiveMessageThread",messages); 
+            //await Clients.Group(groupName).SendAsync("ReceiveMessageThread",messages); 
         }
 
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
-            await RemoveFromMessageGroup();
+
+            var group = await RemoveFromMessageGroup();
+            await Clients.Group(group.Name).SendAsync("UpdateGroup");
+            // await RemoveFromMessageGroup();
             await base.OnDisconnectedAsync(exception);
         }
 
@@ -96,7 +100,7 @@ namespace API.SignalR
         }
 
 
-        private async Task<bool> AddToGroup(string groupName)
+        private async Task<Group> AddToGroup(string groupName)
         {
             var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username");
             var group = await messageRepository.GetMessageGroup(groupName);
@@ -109,17 +113,48 @@ namespace API.SignalR
             }
             group.Connections.Add(connection);
 
-            return await messageRepository.SaveAllAsync();
+            if (await messageRepository.SaveAllAsync()) return group;
+            throw new HubException("Failed to join group");
         }
-        private async Task RemoveFromMessageGroup()
+
+        // private async Task<bool> AddToGroup(string groupName)
+        // {
+        //     var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username");
+        //     var group = await messageRepository.GetMessageGroup(groupName);
+        //     var connection = new Connection { ConnectionId = Context.ConnectionId, Username = username };
+
+        //     if (group == null)
+        //     {
+        //         group = new Group { Name = groupName };
+        //         messageRepository.AddGroup(group);
+        //     }
+        //     group.Connections.Add(connection);
+
+        //     return await messageRepository.SaveAllAsync();
+        // }
+
+        private async Task<Group> RemoveFromMessageGroup()
         {
-            var connection = await messageRepository.GetConnection(Context.ConnectionId);
-            if(connection != null)
+            var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+            var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+            // var connection = await messageRepository.GetConnection(Context.ConnectionId);
+            if (connection != null && group != null)
             {
                 messageRepository.RemoveConnection(connection);
-                await messageRepository.SaveAllAsync();
+                if (await messageRepository.SaveAllAsync()) return group;
+                //await messageRepository.SaveAllAsync();
             }
+            throw new Exception("Faild to remove from group");
         }
+        // private async Task RemoveFromMessageGroup()
+        // {
+        //     var connection = await messageRepository.GetConnection(Context.ConnectionId);
+        //     if(connection != null)
+        //     {
+        //         messageRepository.RemoveConnection(connection);
+        //         await messageRepository.SaveAllAsync();
+        //     }
+        // }
 
         private string GetGroupName(string caller, string? other)
         {
